Refactor for cleanliness
