+++ QRCodeTarot/ViewControllers/CardDetailViewController/CardDetailViewController.swift
@@ -9,41 +9,35 @@ import UIKit
 import TableMVVM
 import CommonUIKitExtensions
 
-class CardDetailViewController: UIViewController {
+class CardDetailViewController: UIViewController, HasPayload {
     
     @IBOutlet var cardImageView: CardImageView!
     @IBOutlet var collapsibleLabel: CollapsableLabelLabel!
     @IBOutlet var evolvedWithText: EvolvedWithText!
 
-    var card: Card! // unit tested.
-
-    static func instantiate(card: Card) -> CardDetailViewController {
-        let detailController: CardDetailViewController = UIStoryboard.vc()!
-        detailController.card = card
-        return detailController
-    }
+    var payload: Card! // unit tested.
 
     override func viewDidLoad() {
         super.viewDidLoad()
         navigationController?.navigationBar.topItem?.title = ""
-        title = card.name.capitalized
+        title = payload.name.capitalized
         let lineCount = UILabel
             .with(width: UIScreen.main.bounds.width - 48)
             .with(lineCount: 0)
-            .with(text: card.desc)
+            .with(text: payload.desc)
             .actualLineCount
         
-        cardImageView.viewModel = card.image ?? .cardDemo
+        cardImageView.viewModel = payload.image ?? .cardDemo
         collapsibleLabel.viewModel = .init(
             labelLabelViewModel: LabelLabel.ViewModel(
                 topText: "Description",
-                bottomText: card.desc,
-                lineCount: lineCount < 6 || card.hasNoEvolutionContent ? 0 : 4
+                bottomText: payload.desc,
+                lineCount: lineCount < 6 || payload.hasNoEvolutionContent ? 0 : 4
             ),
             buttonText: "Read More",
-            buttonIsHidden: lineCount < 6 || card.hasNoEvolutionContent
+            buttonIsHidden: lineCount < 6 || payload.hasNoEvolutionContent
         )
-        evolvedWithText.viewModel = card.evolvedSwitchViewModel
+        evolvedWithText.viewModel = payload.evolvedSwitchViewModel
         view.set(background: BackgroundView.zero.darkShade)
         collapsibleLabel.viewModel.buttonTapped =  { [weak collapsibleLabel] in
             collapsibleLabel?.viewModel.labelLabelViewModel.lineCount = 0
